+++ rcl_lifecycle/src/rcl_lifecycle.c
@@ -50,7 +50,7 @@ rcl_lifecycle_get_zero_initialized_state()
 rcl_ret_t
 rcl_lifecycle_state_init(
   rcl_lifecycle_state_t * state,
-  unsigned int id,
+  uint8_t id,
   const char * label,
   const rcl_allocator_t * allocator)
 {
